Modified filter
Added an "Any" value to the filters. This changes clarifies to the user
what they need to select in order for the filters to retrieve cards that
are of any value for the filter type. There are filter combinations where a section of 0 should return
only one card, but instead, returns a list of cards with any value for the section. This modification also resolves that issue.